Move is_custom_modeling_model
